// src/world/regions/generated/seasia.mask.ts
export const CELL_W = 8;
export const CELL_H = 11;
// ...
export const ANCHORS: Record<string, { x: number; y: number }> = {
  manila: { x: 446, y: 39 },
  m_dock: { x: 415, y: 57 },
  m_spawn: { x: 397, y: 78 },
  school_dh_1: { x: 296, y: 23 },
  school_dh_2: { x: 306, y: 39 },
  school_cw_1: { x: 429, y: 107 },
  school_cw_2: { x: 438, y: 125 },
  school_br_1: { x: 832, y: 508 },
  school_br_2: { x: 890, y: 615 },
  label_malacca: { x: 77, y: 218 },
};